deployment/lambda/requirements.txt
 #Lambda specific dependencies

 # Core dependencies
 boto3
 requests
 python-dotenv

 # AI Assistant dependencies
 openai
 pydantic
 telebot
 python-telegram-bot
